// Lezioni/24.11.2025/vite-project/src/core/models/Items.ts
export interface SkyblockItems {
  success: boolean;
  lastUpdated: number;
  items: Item[];
}

export interface Item {
  material: string;
  durability: any;
  skin?: Skin;
  name: string;
  category?: string;
  tier?: string;
  npc_sell_price?: number;
  id: string;
  salvages?: Salvage[];
  rarity_salvageable?: boolean;
  description?: string;
  item_model?: string;
  stats?: Stats;
  unstackable?: boolean;
  dungeon_item_conversion_cost?: DungeonItemConversionCost;
  upgrade_costs?: UpgradeCost[][];
  museum_data?: MuseumData;
  requirements?: Requirement[];
  color?: string;
  soulbound?: string;
  has_uuid: any;
  gemstone_slots?: GemstoneSlot[];
  glowing?: boolean;
  can_auction?: boolean;
  can_trade?: boolean;
  can_place?: boolean;
  museum?: boolean;
  generator?: string;
  generator_tier?: number;
  furniture?: string;
  item_specific?: ItemSpecific;
  gear_score?: number;
  dungeon_item?: boolean;
  catacombs_requirements?: CatacombsRequirement[];
  can_recombobulate?: boolean;
  salvageable_from_recipe?: boolean;
  motes_sell_price?: number;
  double_tap_to_drop?: boolean;
  enchantments?: Enchantments;
  rift_transferrable?: boolean;
  origin?: string;
  hide_from_viewrecipe_command?: boolean;
  ability_damage_scaling?: number;
  tiered_stats?: TieredStats;
  crystal?: string;
  can_burn_in_furnace?: boolean;
  salvage?: Salvage2;
  serializable?: boolean;
  can_have_attributes?: boolean;
  can_interact?: boolean;
  can_interact_right_click?: boolean;
  private_island?: string;
  can_have_power_scroll?: boolean;
  can_interact_entity?: boolean;
  MINING_FORTUNE?: number;
  sword_type?: string;
  is_upgradeable_without_soulbinding?: boolean;
  recipes?: Recipe[];
  cannot_reforge?: boolean;
  lose_motes_value_on_transfer?: boolean;
  can_have_booster?: boolean;
  prestige?: Prestige;
}

export interface Skin {
  value: string;
  signature?: string;
}

export interface Salvage {
  type: string;
  essence_type?: string;
  amount: number;
  item_id?: string;
}

export interface Stats {
  INTELLIGENCE?: number;
  health_regeneration?: number;
  health?: number;
  fishing_speed?: number;
  BREAKING_POWER?: number;
  DAMAGE?: number;
  MINING_SPEED?: number;
  damage?: number;
  strength?: number;
  DEFENSE?: number;
  HEALTH?: number;
  TRUE_DEFENSE?: number;
  WALK_SPEED?: number;
  mending?: number;
  STRENGTH?: number;
  WEAPON_ABILITY_DAMAGE?: number;
  ABILITY_DAMAGE_PERCENT?: number;
  CRITICAL_DAMAGE?: number;
  SEA_CREATURE_CHANCE?: number;
  defense?: number;
  sea_creature_chance?: number;
  SWING_RANGE?: number;
  CRITICAL_CHANCE?: number;
  critical_damage?: number;
  farming_wisdom?: number;
  fishing_wisdom?: number;
  pressure_resistance?: number;
  fig_fortune?: number;
  mangrove_fortune?: number;
  respiration?: number;
  sweep?: number;
  foraging_fortune?: number;
  FORAGING_FORTUNE?: number;
  SWEEP?: number;
  pull?: number;
  magic_find?: number;
  RIFT_DAMAGE?: number;
  rift_damage?: number;
  rift_intelligence?: number;
  rift_health?: number;
  rift_mana_regen?: number;
  vitality?: number;
  rift_time?: number;
  rift_walk_speed?: number;
  intelligence?: number;
  RIFT_INTELLIGENCE?: number;
  walk_speed?: number;
  weapon_ability_damage?: number;
  RIFT_TIME?: number;
  tracking?: number;
  rift_Time?: number;
  FARMING_FORTUNE?: number;
  farming_fortune?: number;
  bonus_pest_chance?: number;
  wheat_fortune?: number;
  carrot_fortune?: number;
  potato_fortune?: number;
  melon_fortune?: number;
  pumpkin_fortune?: number;
  cocoa_beans_fortune?: number;
  MINING_FORTUNE?: number;
  COLD_RESISTANCE?: number;
  DWARVEN_METAL_FORTUNE?: number;
  combat_wisdom?: number;
  ATTACK_SPEED?: number;
  foraging_wisdom?: number;
  mining_speed?: number;
  mining_fortune?: number;
  treasure_chance?: number;
  FEROCITY?: number;
  HEAT_RESISTANCE?: number;
  GEMSTONE_FORTUNE?: number;
  MAGIC_FIND?: number;
  trophy_fish_chance?: number;
  PRISTINE?: number;
  ORE_FORTUNE?: number;
  PET_LUCK?: number;
  BLOCK_FORTUNE?: number;
  FEAR?: number;
  ferocity?: number;
  double_hook_chance?: number;
  heat_resistance?: number;
  cold_resistance?: number;
}

export interface DungeonItemConversionCost {
  essence_type: string;
  amount: number;
}

export interface UpgradeCost {
  type: string;
  essence_type?: string;
  amount: number;
  item_id?: string;
}

export interface MuseumData {
  type: string;
  armor_set_donation_xp?: ArmorSetDonationXp;
  parent?: Parent;
  game_stage: string;
  donation_xp?: number;
  mapped_item_ids?: string[];
}

export interface ArmorSetDonationXp {
  ICHTHYIC?: number;
  GILLSPLASH?: number;
  FINWAVE?: number;
  FERVOR?: number;
  HOLLOW?: number;
  AURORA?: number;
  CRIMSON?: number;
  TERROR?: number;
  THUNDER?: number;
  ADAPTIVE?: number;
  SHADOW_ASSASSIN?: number;
  ABYSSAL?: number;
  FIG?: number;
  CANOPY?: number;
  SNORKELING?: number;
  PARK?: number;
  MANGROVE?: number;
  END?: number;
  ELEANOR?: number;
  NUTCRACKER?: number;
  SNOW_SUIT?: number;
  LOTUS?: number;
  MELON?: number;
  SQUASH?: number;
  RABBIT?: number;
  SEYMOUR?: number;
  CROPIE?: number;
  BIOHAZARD?: number;
  FERMENTO?: number;
  ARMOR_OF_THE_PACK?: number;
  MASTIFF?: number;
  REAPER?: number;
  REVENANT?: number;
  PRIMORDIAL?: number;
  TARANTULA?: number;
  MONSTER_RAIDER?: number;
  FINAL_DESTINATION?: number;
  REKINDLED_EMBER?: number;
  NECROMANCER_LORD?: number;
  PUMPKIN?: number;
  WISE_WITHER?: number;
  LAVA_SEA_CREATURE?: number;
  MOLTEN?: number;
  ZOMBIE_SOLDIER?: number;
  MERCENARY?: number;
  TITANIUM?: number;
  GOLD_HUNTER?: number;
  SPEEDSTER?: number;
  STRONG_DRAGON?: number;
  LAPIS_ARMOR?: number;
  ZOMBIE_LORD?: number;
  MINER_OUTFIT?: number;
  EMERALD_ARMOR?: number;
  OLD_DRAGON?: number;
  SHIMMERING_LIGHT?: number;
  SPOOKY?: number;
  HEAVY?: number;
  CRIMSON_HUNTER?: number;
  MAGMA_LORD?: number;
  SKELETON_LORD?: number;
  MUSHROOM?: number;
  HOLY_DRAGON?: number;
  FANCY_TUXEDO?: number;
  BRONZE_HUNTER?: number;
  ARMOR_OF_YOG?: number;
  ZOMBIE_COMMANDER?: number;
  TANK_WITHER?: number;
  DIVAN?: number;
  GLACITE?: number;
  GOLEM_ARMOR?: number;
  POWER_WITHER?: number;
  ZOMBIE_KNIGHT?: number;
  FROZEN_BLAZE?: number;
  SKELETON_GRUNT?: number;
  FARM_ARMOR?: number;
  SUPER_HEAVY?: number;
  SORROW?: number;
  ROSETTA?: number;
  LEAFLET?: number;
  GEMSTONE?: number;
  STARLIGHT?: number;
  SQUIRE?: number;
  MITHRIL?: number;
  BLAZE?: number;
  FLAME_BREAKER?: number;
  GOBLIN?: number;
  CACTUS?: number;
  SPONGE?: number;
  CRYSTAL?: number;
  VANQUISHED?: number;
  WITHER?: number;
  ROTTEN?: number;
  CELESTE?: number;
  BERSERKER?: number;
  PARTY?: number;
  SUPERIOR_DRAGON?: number;
  DIAMOND_HUNTER?: number;
  MONSTER_HUNTER?: number;
  BOUNCY?: number;
  FARM_SUIT?: number;
  YOUNG_DRAGON?: number;
  WEREWOLF?: number;
  ZOMBIE?: number;
  DIVER?: number;
  PROTECTOR_DRAGON?: number;
  HARDENED_DIAMOND?: number;
  SILVER_HUNTER?: number;
  FAIRY?: number;
  HEAT?: number;
  SKELETOR?: number;
  UNSTABLE_DRAGON?: number;
  SALMON_NEW?: number;
  SPEED_WITHER?: number;
  WISE_DRAGON?: number;
  GROWTH?: number;
  BAT_PERSON?: number;
  SKELETON_MASTER?: number;
  SKELETON_SOLDIER?: number;
  RAMPART?: number;
  TANK_MINER?: number;
  ELEGANT_TUXEDO?: number;
  CHEAP_TUXEDO?: number;
  SHARK_SCALE?: number;
  CHALLENGER?: number;
  MYTHOS?: number;
  MINERAL?: number;
  GLOSSY_MINERAL?: number;
  FISHERMAN?: number;
  ANGLER?: number;
  BACKWATER?: number;
  PERFECT_TIER_12?: number;
  PERFECT_TIER_13?: number;
  ARACHNE?: number;
}
// ...
export interface Requirement {
  type: string;
  skill?: string;
  level?: number;
  dungeon_type?: string;
  tier?: number;
  slayer_boss_type?: string;
  collection?: string;
  minimum_age_unit?: string;
  minimum_age?: number;
  reward?: string;
  requirements?: Requirement2[];
  lore_index?: number;
  rabbit?: string;
  faction?: string;
  reputation?: number;
  mode?: string;
  kuudra_tier?: string;
}

export interface Requirement2 {
  type: string;
  profile_type?: string;
  tier?: number;
}

export interface GemstoneSlot {
  slot_type: string;
  costs?: Cost[];
}

export interface Cost {
  type: string;
  item_id?: string;
  amount?: number;
  coins?: number;
}

export interface ItemSpecific {
  bundled_item_id?: string;
  bundled_amount?: number;
  speed_on_farming_island?: number;
  extra_pelts?: number;
  permanent_health?: number;
  permanent_crops_farming_fortune?: number;
  default_skin?: DefaultSkin;
  rarity_skins?: RaritySkin[];
  capture_skins?: CaptureSkin[];
  mana_refund?: number;
  cooldown_seconds?: number;
  bonus_heal?: number;
  heal?: number;
  mana_cost?: number;
  bonus_rift_damage_vs_vampire?: number;
  heal_on_hit?: number;
  duration_ticks?: number;
  damage_multiplier?: number;
  rift_time_per_eat?: number;
  motes_percent_per_eat?: number;
  motes_on_join_per_eat?: number;
  stats_on_rift?: StatsOnRift;
  speed_duration_seconds?: number;
  bonus_fishing_speed_per_bucket?: number;
  max_bonus_fishing_speed?: number;
  slow_duration_seconds?: number;
  experience_gained?: number;
  charges?: number;
  consumed_item?: string;
  rift_time_gain?: number;
  ray_texture?: RayTexture;
  memorable_event_key?: string;
  flex_skins?: FlexSkin[];
  regained_rift_time?: number;
  rift_time_regain_on_kill?: number;
  hearts_reduction?: number;
  range_blocks?: number;
  max_players?: number;
  damage_per_player?: number;
  mana_regen_per_player?: number;
  swappables?: Swappables;
  scaling?: Scaling;
  chisel_charges?: number;
  stats?: Stats3;
  rift_stats?: RiftStats;
  duration_seconds?: number;
  range?: number;
  max_other_players?: number;
  particle_color?: ParticleColor;
  speed_boost?: number;
  effect_duration_seconds?: number;
  bonus_experience_chance?: number;
  portal?: Portal;
  max_contacts?: number;
  has_contact_directory?: boolean;
  has_dnd?: boolean;
  max_musicdiscs?: number;
  model?: string;
}

export interface DefaultSkin {
  value: string;
  signature: string;
}

export interface RaritySkin {
  value: string;
  signature: string;
}

export interface CaptureSkin {
  value: string;
  signature: string;
}

export interface StatsOnRift {
  rift_time: number;
  rift_intelligence: number;
  rift_mana_regen: number;
}

export interface RayTexture {
  value: string;
  signature: string;
}

export interface FlexSkin {
  skin: Skin2;
  name: string;
  description: string;
}

export interface Skin2 {
  value: string;
  signature: string;
}

export interface Swappables {
  white: White;
  orange: Orange;
  magenta: Magenta;
  light_blue: LightBlue;
  yellow: Yellow;
  lime: Lime;
  pink: Pink;
  gray: Gray;
  silver: Silver;
  cyan: Cyan;
  purple: Purple;
  blue: Blue;
  brown: Brown;
  green: Green;
  red: Red;
  black: Black;
}
// ...
export interface Scaling {
  tiers: Tier[];
}
// ...
export interface CatacombsRequirement {
  type: string;
  dungeon_type: string;
  level: number;
}

export interface Enchantments {
  aiming?: number;
  ultimate_the_one?: number;
  quantum?: number;
  transylvanian?: number;
  reflection?: number;
  replenish?: number;
  sharpness?: number;
  efficiency?: number;
  power?: number;
  big_brain?: number;
  vicious?: number;
  counter_strike?: number;
  depth_strider?: number;
  respiration?: number;
  aqua_affinity?: number;
  feather_falling?: number;
  small_brain?: number;
  thorns?: number;
  ultimate_chimera?: number;
  pristine?: number;
  rainbow?: number;
}

export interface TieredStats {
  DAMAGE?: number[];
  WEAPON_ABILITY_DAMAGE?: number[];
  INTELLIGENCE?: number[];
  DEFENSE?: number[];
  HEALTH?: number[];
  STRENGTH?: number[];
  WALK_SPEED?: number[];
  CRITICAL_DAMAGE?: number[];
  CRITICAL_CHANCE?: number[];
  ATTACK_SPEED?: number[];
}

export interface Salvage2 {
  item_id: string;
  amount: number;
}

export interface Recipe {
  output: Output;
  ingredient_symbols: IngredientSymbols;
  matrix: string[];
  allow_quick_crafting: boolean;
}

export interface Output {
  item_id: string;
}

export interface IngredientSymbols {
  A: string;
  B: string;
  C: string;
  D: string;
  E: string;
  F: string;
}

export interface Prestige {
  item_id: string;
  costs: Cost2[];
}